feat: Implement CSV import for vests and sales
Adds a new feature to allow users to import RSU releases and sales from CSV files.

Creates a new `/import` endpoint and a corresponding UI page for file uploads. The import form includes a field for the user to specify the stock symbol for vest imports.

Refactors the application to support fractional shares by changing the data type for quantities from `int64` to `float64` across the database, models, and services. This was a necessary change to prevent data loss from the imported CSV files.

Includes unit tests for the CSV parser and integration tests for the import endpoint. Addresses feedback from code review, including fixing a UI bug in the import form and improving test assertions.

+++ internal/db/db.go
@@ -17,7 +17,7 @@ CREATE TABLE IF NOT EXISTS vests (
     id TEXT PRIMARY KEY,              -- Unique identifier for the vest
     date TEXT NOT NULL,               -- Vesting date (YYYY-MM-DD)
     symbol TEXT NOT NULL,             -- Stock ticker symbol (e.g., GOOGL)
-    quantity INTEGER NOT NULL,        -- Number of shares vested
+    quantity REAL NOT NULL,        -- Number of shares vested
     strike_price_cents INTEGER NOT NULL, -- Price per share in USD cents at vest time
     ecb_rate REAL NOT NULL            -- USD to EUR ECB reference rate on the vest date
 );
@@ -27,7 +27,7 @@ CREATE TABLE IF NOT EXISTS vests (
 CREATE TABLE IF NOT EXISTS sales (
     id TEXT PRIMARY KEY,              -- Unique identifier for the sale
     date TEXT NOT NULL,               -- Sale date (YYYY-MM-DD)
-    quantity INTEGER NOT NULL,        -- Total number of shares sold
+    quantity REAL NOT NULL,        -- Total number of shares sold
     price_cents INTEGER NOT NULL,     -- Price per share in USD cents at sale time
     ecb_rate REAL NOT NULL,           -- USD to EUR ECB reference rate on the sale date
     is_settled BOOLEAN NOT NULL DEFAULT 0 -- Flag for CGT calculation status
@@ -38,7 +38,7 @@ CREATE TABLE IF NOT EXISTS sales (
 CREATE TABLE IF NOT EXISTS sale_lots (
     sale_id TEXT NOT NULL,            -- Foreign key to the sales table
     vest_id TEXT NOT NULL,            -- Foreign key to the vests table
-    quantity INTEGER NOT NULL,        -- Number of shares from the vest lot used in this sale
+    quantity REAL NOT NULL,        -- Number of shares from the vest lot used in this sale
     FOREIGN KEY(sale_id) REFERENCES sales(id),
     FOREIGN KEY(vest_id) REFERENCES vests(id),
     PRIMARY KEY (sale_id, vest_id)
@@ -47,7 +47,7 @@ CREATE TABLE IF NOT EXISTS sale_lots (
 CREATE TABLE IF NOT EXISTS settled_sales (
     sale_date TEXT,
     ticker TEXT,
-    num_shares INTEGER,
+    num_shares REAL,
     sale_price_usd INTEGER,
     gain_loss_usd INTEGER,
     book_value_usd INTEGER,
